Add wildcard site alias environments.
If a site alias contains an environment named '*', then this wildcard environment will match for any requested environment name that does not specifically match some other environment in the alias record. The variable ${env-name} will be replaced with the requested environment name. This allows to aliases to be defined in situations where different environments are identical save for field data that varies only by environment name. An example of this is aliases for Pantheon multidev environment.